+++ src/GafferScene/RendererAlgo.cpp
@@ -1725,8 +1725,13 @@ ConstOutputPtr addGafferOutputParameters( const Output *output, const ScenePlug 
 
 	// Add parameters that provide unique identifiers for the render and the
 	// output. This is used by the Catalogue to know when to start a new image
-	// and when to just update an existing one.
-	param->writable()["gaffer:renderID"] = new StringData( fmt::format( "{}:{}", getpid(), reinterpret_cast<uintptr_t>( renderer ) ) );
+	// and when to just update an existing one. If `gaffer:renderID` has been
+	// defined on the Output already then we assume the user knows what they are
+	// doing instead of replacing it - this facilitates "netrender" style workflows.
+	if( !param->member<StringData>( "gaffer:renderID" ) )
+	{
+		param->writable()["gaffer:renderID"] = new StringData( fmt::format( "{}:{}", getpid(), reinterpret_cast<uintptr_t>( renderer ) ) );
+	}
 	param->writable()["gaffer:outputID"] = new StringData( outputID );
 
 	const Node *node = scene->node();