+++ Changes.md
@@ -21,10 +21,17 @@ Breaking Changes
 1.6.x.x (relative to 1.6.2.0)
 =======
 
+Improvements
+------------
+
+- LightEditor, RenderPassEditor, AttributeEditor : Column width adjustments are now preserved when switching tabs.
+
 Fixes
 -----
 
 - LightEditor : Fixed bug that could cause filter evaluation in an invalid context.
+- Outputs : Custom `gaffer:renderID` parameters are now preserved, allowing custom "netrender" style workflows to be implemented.
+- RenderPassEditor : Fixed bug that could cause duplicate tabs to appear when registering custom columns to an existing section via a different group key.
 
 1.6.2.0 (relative to 1.6.1.0)
 =======
@@ -305,10 +312,14 @@ Build
 - Qt.py : Updated to version 1.4.6.
 - USD : Updated to version 25.05.01.
 
-1.5.x.x (relative to 1.5.16.2)
-=======
+1.5.16.x (relative to 1.5.16.2)
+========
 
+Fixes
+-----
 
+- Outputs : Custom `gaffer:renderID` parameters are now preserved, allowing custom "netrender" style workflows to be implemented.
+- RenderPassEditor : Fixed bug that could cause duplicate tabs to appear when registering custom columns to an existing section via a different group key.
 
 1.5.16.2 (relative to 1.5.16.1)
 ========